pyproject.toml
[tool.poetry]
name = "chatapp"
version = "0.1.0"
description = ""
authors = ["Leonid Terletskyi <leo.terletskyi@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.4"
djangorestframework = "^3.14.0"
djoser = "^2.2.0"
channels = {extras = ["daphne"], version = "^4.0.0"}
channels-redis = "^4.1.0"
django-cors-headers = "^4.2.0"
psycopg2-binary = "^2.9.7"
python-dotenv = "^1.0.0"
pillow = "9.5.0"
django-versatileimagefield = "^3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
